+++ Logos/src/utopia/logos/database/access/many/text/word/ManyWordsAccess.scala
@@ -33,7 +33,6 @@ trait ManyWordsAccess extends ManyWordsAccessLike[StoredWord, ManyWordsAccess] w
 	// IMPLEMENTED	--------------------
 	
 	override def factory = WordDbFactory
-	
 	override protected def self = this
 	
 	override def apply(condition: Condition): ManyWordsAccess = ManyWordsAccess(condition)